Update README.md
# Indian-weather-repository
Overview The INDIAN-WEATHER-REPOSITORY is a collection of Jupyter Notebook files related to Indian weather data. It aims to provide insights, visualizations, and analyses of weather patterns across different regions in India.

Features Data Sources: The sources of weather data used in the repository is based on KAGGLE. In whichthe data is update on the daily based. the weather data is also shows airppm, moonphasesand geolocation.